+++ src/main/java/com/b1n_ry/yigd/mixin/CompassItemMixin.java
@@ -3,6 +3,7 @@ package com.b1n_ry.yigd.mixin;
 import com.b1n_ry.yigd.config.YigdConfig;
 import com.b1n_ry.yigd.data.DeathInfoManager;
 import com.b1n_ry.yigd.data.GraveStatus;
+import com.b1n_ry.yigd.util.GraveCompassHelper;
 import net.minecraft.entity.Entity;
 import net.minecraft.item.CompassItem;
 import net.minecraft.item.ItemStack;
@@ -19,11 +20,16 @@ import java.util.UUID;
 public class CompassItemMixin {
     @Inject(method = "inventoryTick", at = @At("TAIL"))
     private void onInventoryTick(ItemStack stack, World world, Entity entity, int slot, boolean selected, CallbackInfo ci) {
-        // This mixin is to make the grave compass disappear when the grave is broken
+        if (world.isClient()) return;
 
+        // This mixin is to make the grave compass disappear when the grave is broken (or redirect to the closest grave)
         if (!YigdConfig.getConfig().extraFeatures.graveCompass.deleteWhenUnlinked) return;
         NbtCompound itemNbt = stack.getNbt();
         if (itemNbt == null) return;
+
+        if (world.getTime() % 1200 == 0) {
+            GraveCompassHelper.updateClosestNbt(world.getRegistryKey(), entity.getBlockPos(), entity.getUuid(), stack);
+        }
 
         if (!itemNbt.contains("linked_grave")) return;